agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aluminatiai"
version = "0.2.2"
description = "GPU energy monitoring agent — per-job cost attribution for AI teams"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.8"
keywords = ["gpu", "monitoring", "energy", "mlops", "cost", "nvidia"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "nvidia-ml-py>=12.0.0",
    "requests>=2.28",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
prometheus    = ["prometheus-client>=0.19"]
secure        = ["cryptography>=42.0"]
observability = ["PyYAML>=6.0"]
all           = ["prometheus-client>=0.19", "cryptography>=42.0", "PyYAML>=6.0"]

[project.scripts]
aluminatiai = "aluminatiai.cli:main"

[project.urls]
Homepage = "https://aluminatiai.com"
Documentation = "https://aluminatiai.com/docs/agent"
Repository = "https://github.com/AgentMulder404/AluminatAI"
"Bug Tracker" = "https://github.com/AgentMulder404/AluminatAI/issues"

# Map the Python package `aluminatiai` to the files in this directory (agent/).
# All sibling directories that have __init__.py become subpackages.
[tool.setuptools]
packages = [
    "aluminatiai",
    "aluminatiai.attribution",
    "aluminatiai.integrations",
    "aluminatiai.schedulers",
    "aluminatiai.efficiency",
]

[tool.setuptools.package-dir]
aluminatiai               = "."
"aluminatiai.attribution"  = "attribution"
"aluminatiai.integrations" = "integrations"
"aluminatiai.schedulers"   = "schedulers"
"aluminatiai.efficiency"   = "efficiency"

[tool.setuptools.package-data]
"aluminatiai" = ["*.txt", "*.md"]
